# APIDOC.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/cart": {
      "get": {
        "tags": [
          "book-controller"
        ],
        "summary": "Get cart total price.",
        "operationId": "getTotalPrice",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseDouble"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "book-controller"
        ],
        "summary": "Add books to shopping cart.",
        "operationId": "addBookToShoppingCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBookToCartReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseShoppingCart"
                }
              }
            }
          }
        }
      }
    },
    "/books": {
      "get": {
        "tags": [
          "book-controller"
        ],
        "summary": "Get available book list.",
        "operationId": "getBookList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListBook"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "book-controller"
        ],
        "summary": "Create new book or add stock to existing book.",
        "operationId": "createOrAddBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBook"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddBookToCartReq": {
        "required": [
          "bookId"
        ],
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BaseResponseShoppingCart": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ShoppingCart"
          }
        }
      },
      "BookInfo": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string"
          }
        }
      },
      "ShoppingCart": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingItem"
            }
          }
        }
      },
      "ShoppingItem": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string"
          },
          "bookInfo": {
            "$ref": "#/components/schemas/BookInfo"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreateBookReq": {
        "required": [
          "author",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string",
            "enum": [
              "FICTION",
              "EDUCATION"
            ]
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BaseResponseBook": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Book"
          }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BaseResponseDouble": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "BaseResponseListBook": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          }
        }
      }
    }
  }
}