# src/main/java/org/zand/model/BookInfo.java
package org.zand.model;

import lombok.*;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class BookInfo {
    private String title;
    private String author;
    private Double price;
    private String category;
}