# src/main/java/org/zand/model/ShoppingItem.java
package org.zand.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ShoppingItem {
    private String bookId;
    private BookInfo bookInfo;
    private int amount;
}